ITS: Change data type of sector/block size variable in the ITS
The commit changes the data type (from uint16_t to uint32_t) of sector
and block size variables in the ITS config structure. The change in data
type allows the support of flash with sector size more than equal to
64KB.

+++ secure_fw/partitions/internal_trusted_storage/flash_fs/its_flash_fs.h
@@ -50,10 +50,10 @@ extern "C" {
 struct its_flash_fs_config_t {
     const void *flash_dev;    /**< Pointer to the flash device */
     uint32_t flash_area_addr; /**< Base address of the flash region */
-    uint16_t sector_size;     /**< Size of the flash device's physical erase
+    uint32_t sector_size;     /**< Size of the flash device's physical erase
                                *   unit
                                */
-    uint16_t block_size;      /**< Size of a logical filesystem erase unit, a
+    uint32_t block_size;      /**< Size of a logical filesystem erase unit, a
                                *   multiple of sector_size.
                                */
     uint16_t num_blocks;      /**< Number of logical erase blocks */
